tests/wpt/tests/long-animation-frame/tentative/loaf-source-location.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Long Animation Frame Timing: source location extraction</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>

<body>
<h1>Long Animation Frame: source location extraction</h1>
<div id="log"></div>
<script>

promise_test(async t => {
  const [entry, script] = await expect_long_frame_with_script(() => {
    requestAnimationFrame(function non_bound_function() {
      busy_wait();
    });
  }, script => script.invoker === "FrameRequestCallback", t);
  assert_equals(script.sourceURL, location.href);
  assert_equals(script.sourceFunctionName, "non_bound_function");
  assert_greater_than(script.sourceCharPosition, 0);
}, "Source location should be extracted from non-bound functions");

promise_test(async t => {
  const [entry, script] = await expect_long_frame_with_script(() => {
    const object = {};
    requestAnimationFrame((function my_bound_function() {
      busy_wait();
    }).bind(object));
  }, script => script.invoker === "FrameRequestCallback", t);
  assert_equals(script.sourceURL, location.href);
  assert_equals(script.sourceFunctionName, "my_bound_function");
  assert_greater_than(script.sourceCharPosition, 0);
}, "Source location should be extracted from bound functions");

promise_test(async t => {
  const [entry, script] = await expect_long_frame_with_script(() => {
    t.step_timeout(function my_timeout() {
      busy_wait();
    });
  }, script => script.invoker === "TimerHandler:setTimeout" && script.sourceURL, t );
  assert_true(script.sourceURL.includes("testharness.js"));
}, "Source location should be extracted for setTimeout");

promise_test(async t => {
  const scriptLocation = new URL("resources/promise-generates-loaf.js", location.href);
  const [entry, script] = await expect_long_frame_with_script(() => {
    const scriptElement = document.createElement("script");
    scriptElement.src = scriptLocation;
    document.body.appendChild(scriptElement);
  }, script => script.invoker === "Window.fetch.then", t);
  assert_true(script.sourceURL.includes("promise-generates-loaf.js"));
}, "Source location should be extracted for promises");

</script>
</body>
